create course from instructor

+++ app/Http/Controllers/AdminController.php
@@ -12,8 +12,11 @@ class AdminController extends Controller
 {
     //
     public function index(){
-   
-        return view('admin.index');
+        $data = array();
+        if(Session::has('loggedInUser')){
+            $data = UserAuth::where('id','=',Session::get('loggedInUser'))->first();
+        }
+        return view('admin.index')->with('data',$data);
     }
     public function applications(){
         $applications = Instructorapplication::get();
@@ -56,8 +59,19 @@ class AdminController extends Controller
      }
      public function delete($id){
         $application = Instructorapplication::find($id);
-        $application->delete();
-        return redirect()->route('applications');
+        if($application){
+            $user = UserAuth::where('id','=',$application->user_id)->first();
+            $user->role=0;
+            $user->save();
+            if($user){
+                $application->delete();
+                return redirect()->route('applications');
+            }
+       
+        }   
      }
- 
+ public function instructors(){
+    $instructors = Instructorapplication::where('status','=','1')->get();
+    return view('admin.instructors')->with('instructors',$instructors);
+ }
 }
